requirements.txt
torch
torchvision
transformers
opencv-python
kornia
numpy
pandas
scikit-learn
Pillow
anomalib
tqdm
